- Refactored Header and Home to keep all navigation and search bars in the same module. - Added some styling to articles.
