+++ NewsReal/client/src/App.js
@@ -9,16 +9,12 @@ import { NewsProvider } from './providers/NewsProvider';
 
 export const App = () => {
   const { isLoggedIn } = useContext(UserProfileContext);
-  const [open, setOpen] = React.useState(false);
-  const handleDrawerChange = () => {
-    setOpen(!open);
-  };
 
   return (
     <Router>
       <NewsProvider>
-        {isLoggedIn ? <Header open={open} handleDrawerChange={handleDrawerChange} /> : ""}
-        <ApplicationViews open={open} handleDrawerChange={handleDrawerChange} />
+        {isLoggedIn ? <Header /> : ""}
+        <ApplicationViews />
       </NewsProvider>
     </Router>
   );
